test: let Jasmine handle promises
* lock in portrait mode to avoid trouble on CI

The cordova-paramedic runs the saucelabs CI tests in portrait mode. For
the Android 7.0 target, it seemed as our locking to landscape interferes
with the Appium driver trying to set the app to portrait orientation.
The result is a failure to connect to the device and finally a timeout.
To avoid that, we are locking the app into portrait mode instead.

* test: verify orientation after locking